move model file to notifications submodule and delete inspection submodule
