+++ src/nwb2bids/_converters/_dandi_utils.py
@@ -2,7 +2,8 @@ import typing
 
 import pydantic
 
-from .._inspection._inspection_result import Category, InspectionResult, Severity
+from nwb2bids.notifications._inspection_result import Category, InspectionResult, Severity
+
 from ..bids_models import DatasetDescription
 from ..bids_models._model_globals import _BIDS_RRID
