+++ src/nwb2bids/_converters/_dataset_converter.py
@@ -6,11 +6,12 @@ import pandas
 import pydantic
 import typing_extensions
 
+from nwb2bids.notifications._inspection_result import Category, InspectionResult, Severity
+
 from ._dandi_utils import get_bids_dataset_description
 from ._run_config import RunConfig
 from ._session_converter import SessionConverter
 from .._converters._base_converter import BaseConverter
-from .._inspection._inspection_result import Category, InspectionResult, Severity
 from ..bids_models import BidsSessionMetadata, DatasetDescription
