+++ src/nwb2bids/_converters/_session_converter.py
@@ -6,10 +6,11 @@ import shutil
 import pydantic
 import typing_extensions
 
+from nwb2bids.notifications._inspection_result import InspectionResult
+
 from ._datalad_utils import _content_is_retrieved
 from ._run_config import RunConfig
 from .._converters._base_converter import BaseConverter
-from .._inspection._inspection_result import InspectionResult
 from .._tools import cache_read_nwb
 from ..bids_models import BidsSessionMetadata
