+++ src/nwb2bids/bids_models/_bids_session_metadata.py
@@ -7,6 +7,8 @@ import pydantic
 import pynwb
 import typing_extensions
 
+from nwb2bids.notifications._inspection_result import Category, DataStandard, InspectionResult, Severity
+
 from ._base_metadata_model import BaseMetadataContainerModel
 from ._channels import ChannelTable
 from ._electrodes import ElectrodeTable
@@ -15,7 +17,6 @@ from ._model_globals import _VALID_ID_REGEX
 from ._participant import Participant
 from ._probes import ProbeTable
 from .._converters._run_config import RunConfig
-from .._inspection._inspection_result import Category, DataStandard, InspectionResult, Severity
 from .._tools import cache_read_nwb
 from ..sanitization import Sanitization
