+++ src/nwb2bids/bids_models/_channels.py
@@ -8,7 +8,8 @@ import pydantic
 import pynwb
 import typing_extensions
 
-from .._inspection._inspection_result import InspectionResult
+from nwb2bids.notifications._inspection_result import InspectionResult
+
 from ..bids_models._base_metadata_model import BaseMetadataContainerModel, BaseMetadataModel
