+++ src/nwb2bids/bids_models/_probes.py
@@ -7,7 +7,8 @@ import pydantic
 import pynwb
 import typing_extensions
 
-from .._inspection._inspection_result import Category, DataStandard, InspectionResult, Severity
+from nwb2bids.notifications._inspection_result import Category, DataStandard, InspectionResult, Severity
+
 from ..bids_models._base_metadata_model import BaseMetadataContainerModel, BaseMetadataModel
