+++ src/nwb2bids/notifications/_inspection_result.py
